Utility function used in test_size_nn_sim() to simulate one set of potential outcomes Returns profits for all possible equal sample sizes

Utility function used in test_size_nn_sim() to simulate one set of potential outcomes Returns profits for all possible equal sample sizes

Usage

one_rep_test_size(n_vals, N, s, mu, sigma, K)

Arguments

n_vals

potential values for n (1:(floor(N/K)-1))

N

deployment population

s

standard deviations of the outcome

mu

means of the priors on the mean response

sigma

standard deviations of the priors on the mean response

K

number of arms (treatments)

Value

A 2-column matrix with values of n in the first column and profits in the second column

Examples

one_rep_test_size(1:(floor(10/2)-1), N=10, s=c(10,10), mu= 20, sigma=10, K=2)
#>      n   profit
#> [1,] 1 249.5363
#> [2,] 2 288.4297
#> [3,] 3 282.3753
#> [4,] 4 306.4467